+++ cpu1/drivers/f2837xd/driverlib/xbar.h
@@ -5,10 +5,8 @@
 // TITLE:  C28x X-BAR driver.
 //
 //###########################################################################
-// $TI Release: F2837xD Support Library v3.11.00.00 $
-// $Release Date: Sun Oct  4 15:55:24 IST 2020 $
 // $Copyright:
-// Copyright (C) 2013-2020 Texas Instruments Incorporated - http://www.ti.com/
+// Copyright (C) 2022 Texas Instruments Incorporated - http://www.ti.com
 //
 // Redistribution and use in source and binary forms, with or without 
 // modification, are permitted provided that the following conditions 
@@ -787,7 +785,7 @@ XBAR_forceOutputLatch(XBAR_OutputNum output)
     // Set the bit that forces the corresponding OUTPUTLATCH bit.
     //
     HWREGH(OUTPUTXBAR_BASE + XBAR_O_OUTPUTLATCHFRC) =
-        0x1U << ((uint16_t)output / 2U);
+        (uint16_t)0x1U << ((uint16_t)output / 2U);
 }
 
 //*****************************************************************************
